+++ java/com/turbovnc/rfb/ServerNameParameter.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2013, 2015, 2018, 2020-2023 D. R. Commander.
+/* Copyright (C) 2013, 2015, 2018, 2020-2024 D. R. Commander.
  *                                           All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
@@ -23,7 +23,7 @@ public final class ServerNameParameter extends StringParameter {
 
   public ServerNameParameter(String name, Params params, boolean isGUI,
                              String desc, String defValue) {
-    super(name, params, isGUI, desc, defValue);
+    super(name, params, isGUI, false, desc, defValue);
     setDefault(defValue);
   }